+++ packages/core/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @ayazmo/core
 
+## 0.6.7
+
+### Patch Changes
+
+- fix config validation and private plugin loading
+
+- Updated dependencies []:
+  - @ayazmo/types@0.6.7
+  - @ayazmo/utils@0.6.7
+
 ## 0.6.6
 
 ### Patch Changes